+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [JEP 223: New Version-String Scheme](https://openjdk
 
 ## Version [11-ea] unreleased
 
-_in progress..._
+### New Features And Enhancements
+- Generate Maven Consumer POM files into `.bach/workspace/deploy/maven` folder
+- New JitPack-based factory in `Link` that supports multi-module projects
+### Bug Fixes
+- Fix running custom build program
 
 ## Version [11.7] released 2020-08-05